Move search parser tests into their own file in the common app
